Refactor GenericReferences and TestCaseProperties
