+++ src/Agoda.Analyzers.Test/AgodaCustom/TestCases/AG0001/AG0001References.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Agoda.Analyzers.Test.Helpers.GenericTestHelpers;
+using System;
 using System.Collections.Generic;
 using System.Web.Mvc;
